Задание 02/user_service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0